requirements.txt
Django==3.0.3
celery==4.4.0
django-environ==0.4.5
selenium==3.141.0
redis==3.4.1
django-celery-beat==1.6.0
pyvirtualdisplay==0.2.5
requests==2.24.0
djangorestframework==3.11.0
pandas==1.1.2
bokeh==2.2.1
